move Workers logic to worker.rs
